+++ samples/snippets/zonal_buckets/zonal_snippets_test.py
@@ -20,8 +20,8 @@ import pytest
 from google.cloud.storage import Client
 import contextlib
 
-from google.cloud.storage._experimental.asyncio.async_grpc_client import AsyncGrpcClient
-from google.cloud.storage._experimental.asyncio.async_appendable_object_writer import (
+from google.cloud.storage.asyncio.async_grpc_client import AsyncGrpcClient
+from google.cloud.storage.asyncio.async_appendable_object_writer import (
     AsyncAppendableObjectWriter,
 )
